+++ packages/camera/camera_web/example/integration_test/camera_test.dart
@@ -7,7 +7,7 @@ import 'dart:ui';
 
 import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:camera_web/src/camera.dart';
-import 'package:camera_web/src/camera_settings.dart';
+import 'package:camera_web/src/camera_service.dart';
 import 'package:camera_web/src/types/types.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
@@ -26,7 +26,7 @@ void main() {
     late MediaDevices mediaDevices;
 
     late MediaStream mediaStream;
-    late CameraSettings cameraSettings;
+    late CameraService cameraService;
 
     setUp(() {
       window = MockWindow();
@@ -36,13 +36,13 @@ void main() {
       when(() => window.navigator).thenReturn(navigator);
       when(() => navigator.mediaDevices).thenReturn(mediaDevices);
 
-      cameraSettings = MockCameraSettings();
+      cameraService = MockCameraService();
 
       final videoElement = getVideoElementWithBlankStream(Size(10, 10));
       mediaStream = videoElement.captureStream();
 
       when(
-        () => cameraSettings.getMediaStreamForOptions(
+        () => cameraService.getMediaStreamForOptions(
           any(),
           cameraId: any(named: 'cameraId'),
         ),
@@ -55,7 +55,7 @@ void main() {
 
     group('initialize', () {
       testWidgets(
-          'calls CameraSettings.getMediaStreamForOptions '
+          'calls CameraService.getMediaStreamForOptions '
           'with provided options', (tester) async {
         final options = CameraOptions(
           video: VideoConstraints(
@@ -67,13 +67,13 @@ void main() {
         final camera = Camera(
           textureId: textureId,
           options: options,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
 
         verify(
-          () => cameraSettings.getMediaStreamForOptions(
+          () => cameraService.getMediaStreamForOptions(
             options,
             cameraId: textureId,
           ),
@@ -90,7 +90,7 @@ void main() {
           options: CameraOptions(
             audio: audioConstraints,
           ),
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -115,7 +115,7 @@ void main() {
           'with correct properties', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -128,7 +128,7 @@ void main() {
       testWidgets('initializes the camera stream', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -138,16 +138,15 @@ void main() {
 
       testWidgets(
           'throws an exception '
-          'when CameraSettings.getMediaStreamForOptions throws',
-          (tester) async {
+          'when CameraService.getMediaStreamForOptions throws', (tester) async {
         final exception = Exception('A media stream exception occured.');
 
-        when(() => cameraSettings.getMediaStreamForOptions(any(),
+        when(() => cameraService.getMediaStreamForOptions(any(),
             cameraId: any(named: 'cameraId'))).thenThrow(exception);
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         expect(
@@ -163,7 +162,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -180,7 +179,7 @@ void main() {
 
       testWidgets(
           'initializes the camera stream '
-          'from CameraSettings.getMediaStreamForOptions '
+          'from CameraService.getMediaStreamForOptions '
           'if it does not exist', (tester) async {
         final options = CameraOptions(
           video: VideoConstraints(
@@ -191,7 +190,7 @@ void main() {
         final camera = Camera(
           textureId: textureId,
           options: options,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -204,7 +203,7 @@ void main() {
 
         // Should be called twice: for initialize and play.
         verify(
-          () => cameraSettings.getMediaStreamForOptions(
+          () => cameraService.getMediaStreamForOptions(
             options,
             cameraId: textureId,
           ),
@@ -215,11 +214,29 @@ void main() {
       });
     });
 
+    group('pause', () {
+      testWidgets('pauses the camera stream', (tester) async {
+        final camera = Camera(
+          textureId: textureId,
+          cameraService: cameraService,
+        );
+
+        await camera.initialize();
+        await camera.play();
+
+        expect(camera.videoElement.paused, isFalse);
+
+        camera.pause();
+
+        expect(camera.videoElement.paused, isTrue);
+      });
+    });
+
     group('stop', () {
       testWidgets('resets the camera stream', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -236,7 +253,7 @@ void main() {
       testWidgets('returns a captured picture', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -272,7 +289,7 @@ void main() {
         testWidgets('if the flash mode is auto', (tester) async {
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )
             ..window = window
             ..stream = videoStream
@@ -307,7 +324,7 @@ void main() {
         testWidgets('if the flash mode is always', (tester) async {
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )
             ..window = window
             ..stream = videoStream
@@ -352,7 +369,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -372,7 +389,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -409,7 +426,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         )
           ..window = window
           ..stream = videoStream;
@@ -437,7 +454,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         )
           ..window = window
           ..stream = videoStream;
@@ -468,7 +485,7 @@ void main() {
 
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         )
           ..window = window
           ..stream = videoStream;
@@ -494,7 +511,7 @@ void main() {
 
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )
             ..window = window
             ..stream = videoStream;
@@ -531,7 +548,7 @@ void main() {
 
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )
             ..window = window
             ..stream = videoStream;
@@ -568,7 +585,7 @@ void main() {
 
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )
             ..window = window
             ..stream = videoStream;
@@ -604,7 +621,7 @@ void main() {
 
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           )..window = window;
 
           expect(
@@ -631,11 +648,11 @@ void main() {
       group('getMaxZoomLevel', () {
         testWidgets(
             'returns maximum '
-            'from CameraSettings.getZoomLevelCapabilityForCamera',
+            'from CameraService.getZoomLevelCapabilityForCamera',
             (tester) async {
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           );
 
           final zoomLevelCapability = ZoomLevelCapability(
@@ -644,12 +661,12 @@ void main() {
             videoTrack: MockMediaStreamTrack(),
           );
 
-          when(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+          when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .thenReturn(zoomLevelCapability);
 
           final maximumZoomLevel = camera.getMaxZoomLevel();
 
-          verify(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+          verify(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .called(1);
 
           expect(
@@ -662,11 +679,11 @@ void main() {
       group('getMinZoomLevel', () {
         testWidgets(
             'returns minimum '
-            'from CameraSettings.getZoomLevelCapabilityForCamera',
+            'from CameraService.getZoomLevelCapabilityForCamera',
             (tester) async {
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           );
 
           final zoomLevelCapability = ZoomLevelCapability(
@@ -675,12 +692,12 @@ void main() {
             videoTrack: MockMediaStreamTrack(),
           );
 
-          when(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+          when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .thenReturn(zoomLevelCapability);
 
           final minimumZoomLevel = camera.getMinZoomLevel();
 
-          verify(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+          verify(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .called(1);
 
           expect(
@@ -693,11 +710,11 @@ void main() {
       group('setZoomLevel', () {
         testWidgets(
             'applies zoom on the video track '
-            'from CameraSettings.getZoomLevelCapabilityForCamera',
+            'from CameraService.getZoomLevelCapabilityForCamera',
             (tester) async {
           final camera = Camera(
             textureId: textureId,
-            cameraSettings: cameraSettings,
+            cameraService: cameraService,
           );
 
           final videoTrack = MockMediaStreamTrack();
@@ -711,7 +728,7 @@ void main() {
           when(() => videoTrack.applyConstraints(any()))
               .thenAnswer((_) async {});
 
-          when(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+          when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
               .thenReturn(zoomLevelCapability);
 
           const zoom = 75.0;
@@ -735,7 +752,7 @@ void main() {
               'when the provided zoom level is below minimum', (tester) async {
             final camera = Camera(
               textureId: textureId,
-              cameraSettings: cameraSettings,
+              cameraService: cameraService,
             );
 
             final zoomLevelCapability = ZoomLevelCapability(
@@ -744,7 +761,7 @@ void main() {
               videoTrack: MockMediaStreamTrack(),
             );
 
-            when(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+            when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
                 .thenReturn(zoomLevelCapability);
 
             expect(
@@ -769,7 +786,7 @@ void main() {
               'when the provided zoom level is below minimum', (tester) async {
             final camera = Camera(
               textureId: textureId,
-              cameraSettings: cameraSettings,
+              cameraService: cameraService,
             );
 
             final zoomLevelCapability = ZoomLevelCapability(
@@ -778,7 +795,7 @@ void main() {
               videoTrack: MockMediaStreamTrack(),
             );
 
-            when(() => cameraSettings.getZoomLevelCapabilityForCamera(camera))
+            when(() => cameraService.getZoomLevelCapabilityForCamera(camera))
                 .thenReturn(zoomLevelCapability);
 
             expect(
@@ -805,7 +822,7 @@ void main() {
       testWidgets('returns a correct view type', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();
@@ -821,7 +838,7 @@ void main() {
       testWidgets('resets the video element\'s source', (tester) async {
         final camera = Camera(
           textureId: textureId,
-          cameraSettings: cameraSettings,
+          cameraService: cameraService,
         );
 
         await camera.initialize();