+++ packages/in_app_purchase/in_app_purchase_android/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Updated Android lint settings.
 
+## 0.1.4+4
+
+* Removed dependency on the `test` package.
+
 ## 0.1.4+3
 
 - Updated installation instructions in README.