+++ packages/quick_actions/quick_actions/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Updated Android lint settings.
 
+## 0.6.0+5
+
+* Support only calling initialize once.
+
 ## 0.6.0+4
 
 * Remove references to the Android V1 embedding.